Final spec porting
